Persistencia de datos + Preparacion estructura app para linkado con back end, creacion api comunicacion Back End + Creación emulador Back end (no compila)

+++ app/src/main/java/com/spreadyourmusic/spreadyourmusic/models/Playlist.kt
@@ -5,7 +5,7 @@ package com.spreadyourmusic.spreadyourmusic.models
  * On 7/03/18.
  */
 // TODO:
-class Playlist(val id: Int, val name: String, val creator: User, val artLocationUri: String, val content: List<Song>) : Recommendation{
+class Playlist(val id: Long, val name: String, val creator: User, val artLocationUri: String, val content: List<Song>) : Recommendation{
     fun getShareLink(): String{
         //TODO: El link devuelto ha de ser el que apunta a la misma playlist desde la interfaz web
         return "https://www.google.es/"
